eBPF.md: Document BPF_PSEUDO_MAP_FD for lddw
The "load immediate" instruction can have a specific meaning for some
values of the source register. If it is set to BPF_PSEUDO_MAP_FD (1),
then the verifier considers the immediate as a file descriptor to a map,
and replaces by a pointer accordingly.

This is supported in the kernel since commit:
0246e64d9a5f ("bpf: handle pseudo BPF_LD_IMM64 insn")