eBPF.md: Document BPF_PSEUDO_FUNC for lddw
When the source register of a "load immediate" instruction is set to 4,
the verifier considers the immediate value as an instruction offset for
a BPF function in the program. Once loaded into the destination
register, this function can be passed to the bpf_for_each_map_elem()
helper function and used as a callback.

This is supported in the kernel since commit:
69c087ba6225 ("bpf: Add bpf_for_each_map_elem() helper")