eBPF.md: Document BPF_PSEUDO_MAP_VALUE for lddw
In addition to BPF_PSEUDO_MAP (1), the source register of a "load
immediate" instruction can be set to BPF_PSEUDO_MAP_VALUE (2). In that
case, the verifier expects: a map file descriptor in the immediate field
of the first half of the 128-bit instruction, and an offset in the
immediate field of the second half of the instruction (remember that
lddw spans over two regular instruction fields, 16 bytes in total). It
uses this offset to load an address residing in the value of the first
entry of the map referended by the file descriptor, at the provided
offset within that value. Currently, only single-entry array maps are
supported. Maps with multiple entries might be supported in the future
by using the offset fields as indices within the map.

This is supported in the kernel since commit:
d8eca5bbb2be ("bpf: implement lookup-free direct value access for maps")