eBPF.md: Document BPF_PSEUDO_CALL for the call instruction
By default, the call instruction calls into a helper function, one of
the specific functions in the kernel that have been created for BPF
programs.

When the source register of a call instruction is set to BPF_PSEUDO_CALL
(1), the call refers to a BPF function instead, for which the offset is
passed in the immediate field.

This is supported in the kernel since commit:
cc8b0b92a169 ("bpf: introduce function calls (function boundaries)")