eBPF.md: Document BPF_PSEUDO_KFUNC_CALL for the call instruction
When the source register of a call instruction is set to
BPF_PSEUDO_KFUNC_CALL (2), then the call refers to a kernel function.
This is different from a helper function, in that the function is not
specific to BPF programs (but it must be explicitly marked as call-able
from BPF programs).

This is supported in the kernel since commit:
e6ac2450d6de ("bpf: Support bpf program calling kernel function")